Add usages dropdown dialog
