+++ src/app/shared/components/grid-usages-dropdown-dialog/grid-usages-dropdown-dialog.component.html
@@ -1,0 +1,14 @@
+<app-connected-dropdown-dialog
+  [title]="'Flytning af systemanvendelse for ' + (organizationName$ | async)"
+  i18n-title
+  dropdownText="Vælg IT system"
+  i18n-dropdownText
+  [data$]="unusedItSystemsInOrganization$"
+  [isLoading$]="loadingUnusedItSystemsInOrganization$"
+  confirmButtonText="Se konsekvenser"
+  i18n-confirmButtonText
+  [nested]="true"
+  (filterChange)="onFilterChange($event)"
+  (save)="onConfirm($event)"
+></app-connected-dropdown-dialog>
+
